mjpeg-opencv-server-python/requirements.txt
opencv-python==4.5.5.62
Pillow==9.2.0
requests=2.28.1
